% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_survey.R
\name{import_survey}
\alias{import_survey}
\title{Import a survey to Qualtrics}
\usage{
import_survey(survey_name = NULL, file_path = NULL)
}
\arguments{
\item{survey_name}{String. The name to be associated with the survey in Qualtrics.}

\item{file_path}{String. Path to file describing survey questions according
to the format specified \link[here]{https://www.qualtrics.com/support/survey-platform/survey-module/survey-tools/import-and-export-surveys/#PreparingASimpleFormatTXTFile}.
Needs to be \code{qsd}, \code{doc}, or \code{txt} format.}
}
\description{
Import a Qualtrics survey via API so that question text can be version
controlled.
}
\examples{
\dontrun{
# Register your Qualtrics credentials if you haven't already
qualtrics_api_credentials(
  api_key = "<YOUR-API-KEY>",
  base_url = "<YOUR-BASE-URL>"
)

# Import a Survey
surveys <- all_surveys("Test Survey", "/survey/path")

}

}
\seealso{
See \url{https://api.qualtrics.com/} for documentation on
the Qualtrics API.
}
